% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badsoRts-package.R
\docType{package}
\name{badsoRts-package}
\alias{badsoRts}
\alias{badsoRts-package}
\title{badsoRts: A Collection Of Really Inefficient Sorting Algorithms}
\description{
Provides various inefficient sorting algorithms for numerical vectors.
}
\author{
\strong{Maintainer}: Maciej Bak \email{wsciekly.maciek@gmail.com} (0000-0003-1361-7301)

}
\keyword{internal}
