% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_functions.R
\name{bogobogosort}
\alias{bogobogosort}
\title{Bogobogosort function}
\usage{
bogobogosort(a)
}
\arguments{
\item{a}{Numeric vector to be sorted}
}
\value{
Numeric vector
}
\description{
Sorting a numeric vector with bogobogosort algorithm.
}
\examples{
bogobogosort(c(1,2,4,3))
}
\keyword{bogobogosort}
