% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_functions.R
\name{bozosort}
\alias{bozosort}
\title{Bozosort function}
\usage{
bozosort(a)
}
\arguments{
\item{a}{Numeric vector to be sorted}
}
\value{
Numeric vector
}
\description{
Sorting a numeric vector with bozosort algorithm.
}
\examples{
bozosort(c(1,2,4,3))
}
\keyword{bozosort}
