% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_functions.R
\name{evilsort}
\alias{evilsort}
\title{Evilsort function}
\usage{
evilsort(a)
}
\arguments{
\item{a}{Numeric vector to be sorted}
}
\value{
Numeric vector
}
\description{
Sorting a numeric vector with evilsort algorithm.
}
\examples{
evilsort(c(1,2,4,3))
}
\author{
Maciej Bak, \email{wsciekly.maciek@gmail.com}
}
\keyword{evilsort}
