% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_functions.R
\name{gorosort}
\alias{gorosort}
\title{Gorosort function}
\usage{
gorosort(a)
}
\arguments{
\item{a}{Numeric vector to be sorted}
}
\value{
Numeric vector
}
\description{
Sorting a numeric vector with gorosort algorithm.
}
\examples{
gorosort(c(1,2,4,3))
}
\keyword{gorosort}
