% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_functions.R
\name{sillysort}
\alias{sillysort}
\title{Silly sort function}
\usage{
sillysort(a)
}
\arguments{
\item{a}{Numeric vector to be sorted}
}
\value{
Numeric vector
}
\description{
Sorting a numeric vector with silly sort algorithm.
}
\examples{
sillysort(c(1,2,4,3))
}
\keyword{sillysort}
