% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_functions.R
\name{slowsort}
\alias{slowsort}
\title{Slowsort function}
\usage{
slowsort(a)
}
\arguments{
\item{a}{Numeric vector to be sorted}
}
\value{
Numeric vector
}
\description{
Sorting a numeric vector with slowsort algorithm.
}
\examples{
slowsort(c(1,2,4,3))
}
\keyword{slowsort}
