% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_functions.R
\name{worstsort}
\alias{worstsort}
\title{Worst function}
\usage{
worstsort(a)
}
\arguments{
\item{a}{Numeric vector to be sorted}
}
\value{
Numeric vector
}
\description{
Sorting a numeric vector with worstsort algorithm.
}
\examples{
worstsort(c(1,2,4,3))
}
\keyword{worstsort}
