% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badsoRts-package.R
\docType{package}
\name{badsoRts-package}
\alias{badsoRts}
\alias{badsoRts-package}
\title{badsoRts: A Collection Of Really Inefficient Sorting Algorithms}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides various inefficient sorting algorithms for numeric vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AngryMaciek/badsoRts}
  \item Report bugs at \url{https://github.com/AngryMaciek/badsoRts/issues}
}

}
\author{
\strong{Maintainer}: Maciej Bak \email{wsciekly.maciek@gmail.com} (0000-0003-1361-7301)

}
\keyword{internal}
